backend/requirements.txt
Flask==3.0.2
pymongo==4.6.1
kafka-python==2.0.2
scikit-learn==1.4.0
xgboost==1.7.6
transformers==4.33.0
numpy==1.26.4
pandas==2.1.4
joblib==1.3.2
matplotlib==3.8.0
gunicorn==21.2.0
python-dotenv==1.0.1
tensorflow==2.16.1